urlencode Jenkinsfile before sending to validation endpoint
The Jenkinsfile contents included in the curl request needs to be URL
encoded in order to be validated correctly. Without this, if the
pipeline contains any character that happens to have special meaning in
URLs, the validation will produce false negative. The minimal
example of such situation would be following pipeline:

```groovy
pipeline {
  stages {
    stage('build') {
      steps {
          sh "cd /app/src && npm install"
      }
    }
  }
}
```

The above would produce incorrect error on the "sh" line because & would
be interpreted as the end of Jenkinsfile.